requirements.txt
openai
python-dotenv
requests
google-genai
